+++ lib/ParamReplacer.py
@@ -1,4 +1,6 @@
 from re import findall
+
+from lib.PathFunctions import ender
 class ParamReplace:
     def __init__(self):
         pass
@@ -7,9 +9,10 @@ class ParamReplace:
         c_counter  = []
         returner_list = []
         counter = 0
-        while counter != len(parameter):
+        length = len(parameter)
+        while counter != length:
             temp = value[counter]
-            for i in range(len(parameter)):
+            for i in range(length):
                 value[counter] = replace_str
                 c_counter.append(parameter[i] + '=' + value[i])
             returner_list.append(c_counter)
@@ -19,35 +22,30 @@ class ParamReplace:
         return returner_list
     
     def only_replacement(self, parameter: list, value: list, replace_str: str, only: list) -> list:
-        c_counter  = []
         returner_list = []
+        c_counter  = []
         counter = 0
-        while counter != len(parameter):
+        length = len(parameter)
+        while counter != length:
             temp = value[counter]
-            for i in range(len(parameter)):
+            for index in range(length):
                 value[counter] = replace_str
-                if parameter[i] in only:
-                    c_counter.append(parameter[i] + '=' + value[i])
+                if parameter[index] in only:
+                    c_counter.append(parameter[index] + '=' + value[index])
             returner_list.append(c_counter)
             value[counter] = temp
             counter += 1
             c_counter = []
         x_counter = []
-        for xxx in returner_list:
-            shit = [yyy.split('=')[-1] for yyy in xxx]
+        for item in returner_list:
+            shit = [yyy.split('=')[-1] for yyy in item]
             if replace_str in shit:
-                x_counter.append(xxx)
+                x_counter.append(item)
         return x_counter
 
 
-    def generate_url(self, half_url: str, xdata: str) -> list:
-        returner_list = []
-        for each in xdata:
-            if half_url[-1] != "?":
-                returner_list.append(half_url + '?' + str("&".join(each)))
-            else:
-                returner_list.append(half_url + str("&".join(each)))
-        return returner_list
+    def generate_url(self, half_url: str, parameters: list) -> list:
+        return [ender(half_url, '?') + '&'.join(parameter) for parameter in parameters]
 
     def expand_parameter(self, query_data: str) -> tuple:
         p,q = [],[]
@@ -58,8 +56,8 @@ class ParamReplace:
             return False,False
         return p,q
 
-    def auto(self, upto_path_url, ppath, payload):
-        apath, bpath = self.expand_parameter(ppath)
-        xpath = self.replacement(apath, bpath, payload)
+    def auto(self, upto_path_url, parameter_to_expand, payload):
+        parameter, value = self.expand_parameter(parameter_to_expand)
+        xpath = self.replacement(parameter, value, payload)
         ypath = self.generate_url(upto_path_url, xpath)
         return ypath